requirements.txt
# Base
pip==18.1
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1

# Customized
click==7.1.2
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
pandas==1.2.3
cython
openpyxl